Finished separating nodes
