+++ dynet/gpu-nodes-arith-unary.cu
@@ -1,3 +1,3 @@
 // This is a dummy file that contains the same content as nodes-unary-arith.cc but compiled
 // on CUDA
-#include "nodes-unary-arith.cc"
+#include "nodes-arith-unary.cc"
