+++ dynet/nodes-conv.cc
@@ -20,77 +20,9 @@ using namespace std;
 
 namespace dynet {
 
+// ************* Filter1DNarrow *************
+
 #ifndef __CUDACC__
-
-string AverageColumns::as_string(const vector<string>& arg_names) const {
-  ostringstream s;
-  s << "average_cols(matrix=" << arg_names[0] << ')';
-  return s.str();
-}
-
-Dim AverageColumns::dim_forward(const vector<Dim>& xs) const {
-  DYNET_ASSERT(xs.size() == 1 || xs.size() == 2, "Failed input count check in AverageColumns");
-  int bd = (xs.size() == 1 ? xs[0].bd : max(xs[0].bd, xs[1].bd));
-  return Dim({xs[0].rows()}, bd);
-}
-
-string FoldRows::as_string(const vector<string>& arg_names) const {
-  ostringstream os;
-  os << "fold_rows(" << arg_names[0] << ", nrows=" << nrows << ')';
-  return os.str();
-}
-
-Dim FoldRows::dim_forward(const vector<Dim>& xs) const {
-  unsigned orows = xs[0].rows() / nrows;
-  if ((orows * nrows != xs[0].rows()) || xs.size() != 1 || xs[0].ndims() > 2) {
-    ostringstream s; s << "Bad input dimensions in FoldRows: " << xs;
-    throw std::invalid_argument(s.str());
-  }
-  return Dim({orows, xs[0].cols()});
-}
-
-/* Deprecated
-string Conv1DNarrow::as_string(const vector<string>& arg_names) const {
-  ostringstream os;
-  os << "conv1d_narrow(" << arg_names[0] << ", f=" << arg_names[1] << ')';
-  return os.str();
-}
-
-Dim Conv1DNarrow::dim_forward(const vector<Dim>& xs) const {
-  if (xs.size() != 2) {
-    ostringstream s; s << "Conv1DNarrow requires two inputs: " << xs;
-    throw std::invalid_argument(s.str());
-  }
-  int ocols = xs[0].cols() - xs[1].cols() + 1;
-  if (xs[0].ndims() != 2 || xs[1].ndims() != 2 ||
-      xs[0].rows() != xs[1].rows() ||
-      ocols < 1) {
-    ostringstream s; s << "Bad input dimensions in Conv1DNarrow: " << xs;
-    throw std::invalid_argument(s.str());
-  }
-  return Dim({xs[0].rows(), (unsigned)ocols});
-}
-
-string Conv1DWide::as_string(const vector<string>& arg_names) const {
-  ostringstream os;
-  os << "conv1d_wide(" << arg_names[0] << ", f=" << arg_names[1] << ')';
-  return os.str();
-}
-
-Dim Conv1DWide::dim_forward(const vector<Dim>& xs) const {
-  if (xs.size() != 2) {
-    ostringstream s; s << "Conv1DWide requires two inputs: " << xs;
-    throw std::invalid_argument(s.str());
-  }
-  unsigned ocols = xs[0].cols() + xs[1].cols() - 1;
-  if (xs[0].ndims() != 2 || xs[1].ndims() != 2 ||
-      xs[0].rows() != xs[1].rows()) {
-    ostringstream s; s << "Bad input dimensions in Conv1DWide: " << xs;
-    throw std::invalid_argument(s.str());
-  }
-  return Dim({xs[0].rows(), ocols});
-}
-*/
 
 string Filter1DNarrow::as_string(const vector<string>& arg_names) const {
   ostringstream os;
@@ -114,143 +46,7 @@ Dim Filter1DNarrow::dim_forward(const vector<Dim>& xs) const {
   return Dim({fids, (unsigned)ocols});
 }
 
-string KMaxPooling::as_string(const vector<string>& arg_names) const {
-  ostringstream os;
-  os << "kmaxpool(" << arg_names[0] << ", k=" << k << ", d=" << pooled_dim << ')';
-  return os.str();
-}
-
-Dim KMaxPooling::dim_forward(const vector<Dim>& xs) const {
-  DYNET_ARG_CHECK(pooled_dim < xs[0].nd,
-                          "Tried to MaxDimension on dimension " << pooled_dim << " bigger than input " << xs[0]);
-  DYNET_ARG_CHECK(xs[0].nd < 4,
-                          "MaxDimension not currently supported for tensors of 4 or more dimensions.");
-  DYNET_ARG_CHECK(k >= 1, "Bad bad k in KMaxPooling: " << k);
-  DYNET_ARG_CHECK(k <= xs[0][pooled_dim], 
-                          "Bad k in KMaxPooling: k = " << k << " bigger than the size of pooled dimension " 
-                          << pooled_dim << " with size = " << xs[0][pooled_dim]);
-  Dim ret(xs[0]);
-  ret.set(pooled_dim, k);
-  return ret;
-}
-
-size_t KMaxPooling::aux_storage_size() const {
-  // map of where the entries in f(x) go to entries in x
-  return sizeof(Eigen::DenseIndex) * dim.size();
-}
-
-string SumDimension::as_string(const vector<string>& arg_names) const {
-  ostringstream s;
-  s << "sum_dim(matrix=" << arg_names[0] << ',' << dimension << '}';
-  return s.str();
-}
-
-Dim SumDimension::dim_forward(const vector<Dim>& xs) const {
-  DYNET_ASSERT(xs.size() == 1, "Failed input count check in SumDimension");
-  Dim ret(xs[0]);
-  ret.delete_dim(dimension);
-  return ret;
-}
 #endif
-
-template<class MyDevice>
-void AverageColumns::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 1, "Failed input count check in AverageColumns");
-  unsigned cols = xs[0]->d.cols();
-#ifdef __CUDACC__
-  // The reduction used on CPU is better, but not implemented in GPU
-  fx.t<1>().device(*dev.edevice) = xs[0]->t<2>().chip<1>(0);
-  for(unsigned i = 1; i < cols; ++i)
-    fx.t<1>().device(*dev.edevice) += xs[0]->t<2>().chip<1>(i);
-  fx.t<1>().device(*dev.edevice) = fx.t<1>() / (float)cols;
-#else
-  const Eigen::array<Eigen::DenseIndex, 1> reduction_axis = {1};
-  fx.t<1>().device(*dev.edevice) = xs[0]->t<2>().sum(reduction_axis) / (float)cols;
-#endif
-}
-
-template<class MyDevice>
-void AverageColumns::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  const Eigen::array<Eigen::DenseIndex, 2> broadcasts = {1, xs[0]->d[1]};
-  dEdxi.t<2>().device(*dev.edevice) += (dEdf.t<2>() / (float)xs[0]->d[1]).broadcast(broadcasts);
-}
-DYNET_NODE_INST_DEV_IMPL(AverageColumns)
-
-/* Deprecated
-template<class MyDevice>
-void Conv1DNarrow::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  const unsigned ycols = dim.cols();
-  const unsigned fcols = xs[1]->d.cols();
-  for (unsigned j = 0; j < ycols; ++j) {
-    fx.t<2>().chip<1>(j).device(*dev.edevice) = xs[0]->t<2>().chip<1>(j) * xs[1]->t<2>().chip<1>(0);
-    for (unsigned k = 1; k < fcols; ++k)
-      fx.t<2>().chip<1>(j).device(*dev.edevice) += xs[0]->t<2>().chip<1>(j+k) * xs[1]->t<2>().chip<1>(k);
-  }
-  // TODO: This following version without chip is better, but for some reason dimensions don't match.
-  // Eigen::array<ptrdiff_t, 1> dims; dims[0] = 1;
-  // fx.t<2>().device(*dev.edevice) = xs[0]->t<2>().convolve(xs[1]->t<2>(), dims);
-}
-
-template<class MyDevice>
-void Conv1DNarrow::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  DYNET_ASSERT(i < 2, "Failed input count check in Conv1DNarrow");
-  const unsigned ycols = dim.cols();
-  const unsigned fcols = xs[1]->d.cols();
-  // TODO: Can this be done with a kernel and without using chip?
-  if (i == 0) { // derivative wrt input x
-    for (unsigned j = 0; j < ycols; ++j)
-      for (unsigned k = 0; k < fcols; ++k)
-        dEdxi.t<2>().chip<1>(j+k).device(*dev.edevice) += xs[1]->t<2>().chip<1>(k) * dEdf.t<2>().chip<1>(j);
-  } else { // derivative wrt filter f
-    for (unsigned j = 0; j < ycols; ++j)
-      for (unsigned k = 0; k < fcols; ++k)
-        dEdxi.t<2>().chip<1>(k).device(*dev.edevice) += xs[0]->t<2>().chip<1>(j+k) * dEdf.t<2>().chip<1>(j);
-  }
-}
-DYNET_NODE_INST_DEV_IMPL(Conv1DNarrow)
-
-template<class MyDevice>
-void Conv1DWide::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  TensorTools::zero(fx);
-  const unsigned xcols = xs[0]->d.cols();
-  const unsigned fcols = xs[1]->d.cols();
-  for (unsigned j = 0; j < xcols; ++j)
-    for (unsigned k = 0; k < fcols; ++k)
-      fx.t<2>().chip<1>(j+k).device(*dev.edevice) += xs[1]->t<2>().chip<1>(k) * xs[0]->t<2>().chip<1>(j);
-}
-
-
-template<class MyDevice>
-void Conv1DWide::backward_dev_impl(const MyDevice & dev,
-                             const vector<const Tensor*>& xs,
-                             const Tensor& fx,
-                             const Tensor& dEdf,
-                             unsigned i,
-                             Tensor& dEdxi) const {
-  const unsigned xcols = xs[0]->d.cols();
-  const unsigned fcols = xs[1]->d.cols();
-  if (i == 0) { // derivative wrt input x
-    for (unsigned j = 0; j < xcols; ++j)
-      for (unsigned k = 0; k < fcols; ++k)
-        dEdxi.t<2>().chip<1>(j).device(*dev.edevice) += xs[1]->t<2>().chip<1>(k) * dEdf.t<2>().chip<1>(j + k);
-  } else { // derivative wrt filter f
-    for (unsigned j = 0; j < xcols; ++j)
-      for (unsigned k = 0; k < fcols; ++k)
-        dEdxi.t<2>().chip<1>(k).device(*dev.edevice) += xs[0]->t<2>().chip<1>(j) * dEdf.t<2>().chip<1>(j + k);
-  }
-}
-DYNET_NODE_INST_DEV_IMPL(Conv1DWide)
-*/
 
 template<class MyDevice>
 void Filter1DNarrow::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
@@ -313,6 +109,26 @@ void Filter1DNarrow::backward_dev_impl(const MyDevice & dev,
 }
 DYNET_NODE_INST_DEV_IMPL(Filter1DNarrow)
 
+// ************* FoldRows *************
+
+#ifndef __CUDACC__
+
+string FoldRows::as_string(const vector<string>& arg_names) const {
+  ostringstream os;
+  os << "fold_rows(" << arg_names[0] << ", nrows=" << nrows << ')';
+  return os.str();
+}
+
+Dim FoldRows::dim_forward(const vector<Dim>& xs) const {
+  unsigned orows = xs[0].rows() / nrows;
+  if ((orows * nrows != xs[0].rows()) || xs.size() != 1 || xs[0].ndims() > 2) {
+    ostringstream s; s << "Bad input dimensions in FoldRows: " << xs;
+    throw std::invalid_argument(s.str());
+  }
+  return Dim({orows, xs[0].cols()});
+}
+
+#endif
 
 template<class MyDevice>
 void FoldRows::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
@@ -342,6 +158,37 @@ void FoldRows::backward_dev_impl(const MyDevice & dev,
   //     dEdxi.tb<2>().chip<0>(i * nrows + j) += d.tb<2>().chip<0>(i);
 }
 DYNET_NODE_INST_DEV_IMPL(FoldRows)
+
+// ************* KMaxPooling *************
+
+#ifndef __CUDACC__
+
+string KMaxPooling::as_string(const vector<string>& arg_names) const {
+  ostringstream os;
+  os << "kmaxpool(" << arg_names[0] << ", k=" << k << ", d=" << pooled_dim << ')';
+  return os.str();
+}
+
+Dim KMaxPooling::dim_forward(const vector<Dim>& xs) const {
+  DYNET_ARG_CHECK(pooled_dim < xs[0].nd,
+                          "Tried to MaxDimension on dimension " << pooled_dim << " bigger than input " << xs[0]);
+  DYNET_ARG_CHECK(xs[0].nd < 4,
+                          "MaxDimension not currently supported for tensors of 4 or more dimensions.");
+  DYNET_ARG_CHECK(k >= 1, "Bad bad k in KMaxPooling: " << k);
+  DYNET_ARG_CHECK(k <= xs[0][pooled_dim], 
+                          "Bad k in KMaxPooling: k = " << k << " bigger than the size of pooled dimension " 
+                          << pooled_dim << " with size = " << xs[0][pooled_dim]);
+  Dim ret(xs[0]);
+  ret.set(pooled_dim, k);
+  return ret;
+}
+
+size_t KMaxPooling::aux_storage_size() const {
+  // map of where the entries in f(x) go to entries in x
+  return sizeof(Eigen::DenseIndex) * dim.size();
+}
+
+#endif
 
 template<class MyDevice>
 void KMaxPooling::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
@@ -429,25 +276,59 @@ void KMaxPooling::backward_dev_impl(const MyDevice & dev,
 }
 DYNET_NODE_INST_DEV_IMPL(KMaxPooling)
 
+// ************* KMHNgram *************
+
+#ifndef __CUDACC__
+
+string KMHNGram::as_string(const vector<string>& arg_names) const {
+  ostringstream s;
+  s << "kmh-ngram(" << arg_names[0] << ')';
+  return s.str();
+}
+
+Dim KMHNGram::dim_forward(const vector<Dim>& xs) const {
+  DYNET_ARG_CHECK(xs[0].ndims() == 2, "Bad input dimensions in KMHNGram: " << xs);
+  const unsigned new_cols = xs[0].cols() - n + 1;
+  DYNET_ARG_CHECK(new_cols >= 1, "Bad input dimensions in KMHNGram: " << xs);
+  return Dim({xs[0][0], new_cols});
+}
+
+#endif
+
 template<class MyDevice>
-void SumDimension::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
-  DYNET_ASSERT(xs.size() == 1, "Failed input count check in SumDimension");
-  Eigen::array<int, 1> reduction_axis = {(int)dimension};
-  fx.t<1>().device(*dev.edevice) = xs[0]->t<2>().sum(reduction_axis);
+void KMHNGram::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
+#ifdef __CUDACC__
+  DYNET_RUNTIME_ERR("KMHNGram not implemented for CUDA");
+#else
+  auto x = **xs[0];
+  const int new_cols = x.cols() - n + 1;
+  DYNET_ASSERT(new_cols > 0, "Failed dimension check in KMHNGram");
+  auto res = *fx;
+  res.setZero();
+  for (int j = 0; j < new_cols; ++j) {
+    auto c_j = res.col(j);
+    for (unsigned k = 0; k < n; ++k)
+      c_j += x.col(j + k);
+  }
+#endif
 }
 
 template<class MyDevice>
-void SumDimension::backward_dev_impl(const MyDevice & dev,
+void KMHNGram::backward_dev_impl(const MyDevice & dev,
                              const vector<const Tensor*>& xs,
                              const Tensor& fx,
                              const Tensor& dEdf,
                              unsigned i,
                              Tensor& dEdxi) const {
-  // TODO: limit to 3-dimensional tensor is arbitrary
-  Eigen::array<int, 4> bcast = {1,1,1,1}; bcast[dimension] = dEdxi.d[dimension];
-  Eigen::array<int, 4> morph = {(int)dEdxi.d[0],(int)dEdxi.d[1],(int)dEdxi.d[2],(int)dEdxi.d.bd}; morph[dimension] = 1;
-  dEdxi.tb<3>().device(*dev.edevice) += dEdf.tb<3>().reshape(morph).broadcast(bcast);
+#ifdef __CUDACC__
+  DYNET_RUNTIME_ERR("KMHNGram not implemented for CUDA");
+#else
+  const int c = dEdf.d.cols();
+  for (int j = 0; j < c; ++j)
+    for (unsigned k = 0; k < n; ++k)
+      (*dEdxi).col(j+k) += (*dEdf).col(j);
+#endif
 }
-DYNET_NODE_INST_DEV_IMPL(SumDimension)
+DYNET_NODE_INST_DEV_IMPL(KMHNGram)
 
 } // namespace dynet
